move kernel and try to fix autoray issue
