+++ pennylane/math/utils.py
@@ -636,7 +636,45 @@ def _reduced_row_echelon(binary_matrix):
             currcol[irow] = 0
 
             # XOR the right hand side of the pivot row irow with all of the other rows
-            rref_mat[:, icol:] ^= _np.outer(currcol, rpvt_cols)
+            rref_mat[:, icol:] ^= np.outer(currcol, rpvt_cols)
             icol += 1
 
     return rref_mat.astype(int)
+
+
+def _kernel(binary_matrix):
+    r"""Computes the kernel of a binary matrix on the binary finite field :math:`\mathbb{Z}_2`.
+
+    Args:
+        binary_matrix (array[int]): binary matrix representation of the Hamiltonian
+    Returns:
+        array[int]: nullspace of the `binary_matrix` where each row corresponds to a
+        basis vector in the nullspace
+
+    **Example**
+
+    >>> binary_matrix = np.array([[1, 0, 0, 0, 0, 1, 0, 0],
+    ...                          [0, 0, 1, 1, 1, 1, 1, 1],
+    ...                          [0, 0, 0, 1, 1, 0, 0, 1]])
+    >>> _kernel(binary_matrix)
+    array([[0, 1, 0, 0, 0, 0, 0, 0],
+           [0, 0, 1, 1, 1, 0, 0, 0],
+           [1, 0, 1, 0, 0, 1, 0, 0],
+           [0, 0, 1, 0, 0, 0, 1, 0],
+           [0, 0, 1, 1, 0, 0, 0, 1]])
+    """
+    # Get the columns with and without pivots
+    pivots = (binary_matrix.T != 0).argmax(axis=0)
+    nonpivots = np.setdiff1d(range(len(binary_matrix[0])), pivots)
+
+    # Initialize the nullspace
+    null_vector = np.zeros((binary_matrix.shape[1], len(nonpivots)), dtype=int)
+    null_vector[nonpivots, np.arange(len(nonpivots))] = 1
+
+    # Fill up the nullspace vectors from the binary matrix
+    null_vector_indices = np.ix_(pivots, np.arange(len(nonpivots)))
+    binary_vector_indices = np.ix_(np.arange(len(pivots)), nonpivots)
+    null_vector[null_vector_indices] = -binary_matrix[binary_vector_indices] % 2
+
+    nullspace = null_vector.T
+    return nullspace
